+++ src/main/java/com/gamemoonchul/GamemoonchulApplication.java
@@ -2,7 +2,9 @@ package com.gamemoonchul;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class GamemoonchulApplication {
