+++ src/main/java/com/gamemoonchul/config/RestTemplateConfig.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.common.config;
+package com.gamemoonchul.config;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
