+++ src/main/java/com/gamemoonchul/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.common.config;
+package com.gamemoonchul.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
