+++ src/main/java/com/gamemoonchul/config/jwt/JwtStatus.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.infrastructure.jwt;
+package com.gamemoonchul.config.jwt;
 
 import com.gamemoonchul.common.status.ApiStatusIfs;
 import lombok.AllArgsConstructor;
@@ -12,7 +12,8 @@ import lombok.Getter;
 public enum JwtStatus implements ApiStatusIfs {
   NOT_VALID_TOKEN(11000, "유효하지 않은 토큰입니다."),
   EXPIRED_TOKEN(11001, "만료된 토큰입니다."),
-  SIGNATURE_NOT_MATCH(11002, "토큰의 서명이 일치하지 않습니다."),
+  SIGNATURE_NOT_MATCH(
+      11002, "토큰의 서명이 일치하지 않습니다."),
   NULL_POINT(11003, "토큰이 존재하지 않습니다."),
   ;
