+++ src/main/java/com/gamemoonchul/config/jwt/TokenDto.java
@@ -1,0 +1,11 @@
+package com.gamemoonchul.config.jwt;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class TokenDto {
+  private String accessToken;
+  private String refreshToken;
+}
