+++ src/main/java/com/gamemoonchul/config/oauth/HttpCookieOAuth2AuthorizationRequestRepository.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.infrastructure.oauth;
+package com.gamemoonchul.config.oauth;
 
 import com.gamemoonchul.common.util.CookieUtils;
 import jakarta.servlet.http.HttpServletRequest;
