+++ src/main/java/com/gamemoonchul/config/oauth/user/GoogleOAuth2UserUnlink.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.infrastructure.oauth.user;
+package com.gamemoonchul.config.oauth.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
