+++ src/main/java/com/gamemoonchul/config/oauth/user/OAuth2Provider.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.infrastructure.oauth.user;
+package com.gamemoonchul.config.oauth.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
