+++ src/main/java/com/gamemoonchul/config/oauth/user/OAuth2UserUnlink.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.infrastructure.oauth.user;
+package com.gamemoonchul.config.oauth.user;
 
 public interface OAuth2UserUnlink {
     void unlink(String accessToken);
