+++ src/main/java/com/gamemoonchul/config/oauth/user/OAuth2UserUnlinkManager.java
@@ -1,7 +1,7 @@
-package com.gamemoonchul.infrastructure.oauth.user;
+package com.gamemoonchul.config.oauth.user;
 
 import com.gamemoonchul.common.exception.ApiException;
-import com.gamemoonchul.infrastructure.oauth.Oauth2Status;
+import com.gamemoonchul.config.oauth.Oauth2Status;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
