+++ src/main/java/com/gamemoonchul/domain/status/MemberStatus.java
@@ -1,0 +1,18 @@
+package com.gamemoonchul.domain.status;
+
+import com.gamemoonchul.common.status.ApiStatusIfs;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * 회원 관련 상태 코드 2000 ~ 2999
+ */
+@Getter
+@AllArgsConstructor
+public enum MemberStatus implements ApiStatusIfs {
+  MEMBER_NOT_FOUND(2404, "회원을 찾을 수 없습니다."),
+  ;
+
+  private final Integer statusCode;
+  private final String message;
+}
