--- src/main/java/com/gamemoonchul/infrastructure/oauth/service/CustomOAuth2UserService.java
@@ -1,55 +1,0 @@
-package com.gamemoonchul.infrastructure.oauth.service;
-
-import com.gamemoonchul.common.exception.ApiException;
-import com.gamemoonchul.infrastructure.oauth.OAuth2UserPrincipal;
-import com.gamemoonchul.infrastructure.oauth.Oauth2Status;
-import com.gamemoonchul.infrastructure.oauth.user.OAuth2UserInfo;
-import com.gamemoonchul.infrastructure.oauth.user.OAuth2UserInfoFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.InternalAuthenticationServiceException;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-
-@RequiredArgsConstructor
-@Service
-public class CustomOAuth2UserService extends DefaultOAuth2UserService {
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest oAuth2UserRequest) throws OAuth2AuthenticationException {
-
-        OAuth2User oAuth2User = super.loadUser(oAuth2UserRequest);
-
-        try {
-            return processOAuth2User(oAuth2UserRequest, oAuth2User);
-        } catch (AuthenticationException ex) {
-            // Throwing an instance of AuthenticationException will trigger the OAuth2AuthenticationFailureHandler
-            throw ex;
-        } catch (Exception ex) {
-            throw new ApiException(Oauth2Status.LOGIN_FAILED);
-        }
-    }
-
-    private OAuth2User processOAuth2User(OAuth2UserRequest userRequest, OAuth2User oAuth2User) {
-
-        String registrationId = userRequest.getClientRegistration()
-                .getRegistrationId();
-
-        String accessToken = userRequest.getAccessToken().getTokenValue();
-
-        OAuth2UserInfo oAuth2UserInfo = OAuth2UserInfoFactory.getOAuth2UserInfo(registrationId,
-                accessToken,
-                oAuth2User.getAttributes());
-
-        // OAuth2UserInfo field value validation
-        if (!StringUtils.hasText(oAuth2UserInfo.getEmail())) {
-            throw new ApiException(Oauth2Status.NOT_FOUND_EMAIL);
-        }
-
-        return new OAuth2UserPrincipal(oAuth2UserInfo);
-    }
-}
