--- src/main/resources/application.yaml
@@ -1,5 +1,0 @@
-spring:
-  security:
-    user:
-      name: user
-      password: 1234
